a zip script

+++ zip.sh
@@ -1,2 +1,2 @@
-rm -r lib/__pycache__/
+[ -d lib/__pycache__ ] && rm -r lib/__pycache__/
 zip -r ai27_ppompean_jplacett.zip main.py README.md lib/
